фикс категорий и подкатегорий

+++ src/components/ProductList.vue
@@ -1,28 +1,40 @@
 <template>
-  <VaInput
-      v-model="searchQuery"
-      class="ml-4"
-      placeholder="Поиск"
-      clearable
-  />
-  <VaSelect
-      class="ml-4"
-      :options="sortOptions"
-      v-model="selectedSortOption"
-      placement="right"
-  />
-  <VaSelect
-      class="ml-4"
-      :options="sortType"
-      v-model="selectedSortType"
-      placement="right"
-  />
-  <VaButton
-      preset="secondary"
-      @click="showDeleteConfirmation()"
-  >
-    Удалить все
-  </VaButton>
+  <div class="container">
+    <div class="input-container">
+      <VaInput
+          v-model="searchQuery"
+          class="ml-4"
+          placeholder="Поиск"
+          clearable
+      />
+      <VaSelect
+          class="ml-4"
+          :options="sortOptions"
+          v-model="selectedSortOption"
+          placement="right"
+      />
+      <VaSelect
+          class="ml-4"
+          :options="sortType"
+          v-model="selectedSortType"
+          placement="right"
+      />
+    </div>
+    <div class="btn-container">
+      <VaButton
+          preset="secondary"
+          @click="showDeleteConfirmation()"
+      >
+        Удалить все
+      </VaButton>
+      <VaButton
+          preset="secondary"
+          @click="addCategory(null)"
+      >
+        Добавить категорию
+      </VaButton>
+    </div>
+  </div>
   <div class="container ml-4">
     <VaTreeView
         :key="searchQuery"
@@ -30,6 +42,7 @@
         text-by="name"
         :filter="searchQuery"
         expand-all
+        :expanded="Array(searchQuery)"
     >
       <template #content="{ id, name, type }">
         <div class="tree-node-row">
@@ -51,17 +64,22 @@
                   <VaMenuItem @click="addProduct(id)">Добавить продукт</VaMenuItem>
                 </div>
                 <VaMenuItem @click="editItem(id, type)">Редактировать</VaMenuItem>
-                <VaMenuItem @click="showDeleteConfirmation(id)">Удалить</VaMenuItem>
+                <VaMenuItem @click="showDeleteConfirmation(id, type)">Удалить</VaMenuItem>
               </div>
 
             </VaMenu>
           </div>
         </div>
       </template>
-
-
     </VaTreeView>
   </div>
+  <add-category-modal
+      ref="addCategoryModal"
+      :editableCategory="editable"
+      :parentCategoryId="parentCategoryId"
+      :isEdit="isEdit"
+      @close="()=> {isEdit = false, loadData()}"
+  />
 
   <confirm-modal
       ref="confirmModal"
@@ -79,10 +97,13 @@ import {buildTree} from "@/utils/treeBuilder.ts";
 import type {IProduct} from "@/models/product.model.ts";
 import type {ICategory} from "@/models/category.model.ts";
 import type {TreeNode} from "@/models/tree-node.model.ts";
+import AddCategoryModal from "@/components/modals/AddCategoryModal.vue";
+
 
 const categories = ref<ICategory[]>([]);
 const products = ref<IProduct[]>([]);
 const confirmModal = ref<InstanceType<typeof ConfirmModal> | null>(null);
+const addCategoryModal = ref<InstanceType<typeof AddCategoryModal> | null>(null)
 const toast = useToast();
 const productIdToDelete = ref<number | null>(null);
 const selectedSortOption = ref("По умолчанию");
@@ -90,43 +111,70 @@ const selectedSortType = ref("По убыванию");
 const searchQuery = ref("");
 const baseTree = ref<TreeNode[]>([]);
 const displayTree = ref<TreeNode[]>([]);
-
+const deleteType = ref<'product' | 'category' | null>(null);
+const editable = ref<ICategory | null>(null)
+const isEdit = ref<boolean>(false)
+const parentCategoryId = ref<number | null>(null);
 
 const sortOptions = ["По умолчанию", "По наименованию"];
 const sortType = ["По убыванию", "По возрастанию"];
-function addCategory(){
+
+function addCategory(parentId: number | null) {
+  isEdit.value = false;
+  editable.value = null;
+  parentCategoryId.value = parentId;
+  addCategoryModal.value?.show();
 
 }
-function editItem(id: number){
 
+function editItem(id: number) {
+  isEdit.value = true
+  // editable.value = JSON.parse(JSON.stringify());
+  addCategoryModal.value?.show();
 }
-function addProduct(id: number){
 
+function addProduct(id: number) {
+  isEdit.value = false
+  // editable.value = JSON.parse(JSON.stringify());
+  addCategoryModal.value?.show();
 }
 
-function showDeleteConfirmation(id?: number) {
-  if (id) {
+function showDeleteConfirmation(id?: number, type?: 'product' | 'category') {
+  if (id && type) {
     productIdToDelete.value = id;
+    deleteType.value = type;
+  } else {
+    productIdToDelete.value = null;
+    deleteType.value = null;
   }
   confirmModal.value?.show();
 }
 
+
 async function confirmDelete() {
   try {
-    if (productIdToDelete.value !== null) {
-      await dbService.deleteOneProduct(productIdToDelete.value);
-      toast.init({message: "Товар удалён", color: "success"});
+    if (productIdToDelete.value !== null && deleteType.value !== null) {
+      if (deleteType.value === 'product') {
+        await dbService.deleteOneProduct(productIdToDelete.value);
+        toast.init({message: "Товар удалён", color: "success"});
+      } else if (deleteType.value === 'category') {
+        await dbService.deleteCategory(productIdToDelete.value);
+        toast.init({message: "Категория удалена", color: "success"});
+      }
     } else {
       await dbService.deleteAllProducts();
       toast.init({message: "Товары удалены", color: "success"});
     }
+
     await loadData();
   } catch {
     toast.init({message: "Ошибка при удалении", color: "danger"});
   } finally {
     productIdToDelete.value = null;
+    deleteType.value = null;
   }
 }
+
 
 function sortTreeRecursively(nodes: TreeNode[], sortBy: string, sortOrder: string): TreeNode[] {
   const sorted = [...nodes].sort((a, b) => {
@@ -163,10 +211,31 @@ onMounted(loadData);
 <style lang="scss" scoped>
 .container {
   margin: 10px;
+  display: flex;
+  gap: 10px;
+
+  .input-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 5px;
+
+    .va-input {
+      height: min-content;
+    }
+  }
+
+  .btn-container {
+    display: flex;
+
+    button {
+      width: 100%;
+    }
+  }
 
   .va-list-item {
     font-size: 20px;
   }
+
   .tree-node-row {
     display: flex;
     align-items: center;
